requirements.txt
Django==4.2.9
djangorestframework==3.14.0
mysqlclient==2.2.4
django-cors-headers==4.3.1
python-dotenv==1.0.0
drf-spectacular==0.27.1
drf-nested-routers==0.93.4